[Design] #26 - 편지 내용 페이지 UI 수정
