+++ src/App.js
@@ -12,7 +12,6 @@ import Score1Page from './pages/testScorePage/Score1Page'
 import Score2Page from './pages/testScorePage/Score2Page'
 import Score3Page from './pages/testScorePage/Score3Page'
 import TestPage from './pages/testPage/TestPage'
-import LetterReadPage from './pages/letterPage/LetterReadPage'
 import LetterCreatePage from './pages/letterPage/LetterCreatePage'
 import RankingPage from './pages/testScorePage/RankingPage'
 import LetterTreePage from './pages/letterPage/LetterTreePage'
@@ -34,7 +33,6 @@ const App = () => {
                 <Route path ='/score2' element={<Score2Page/>} />
                 <Route path ='/score3' element={<Score3Page/>} />
                 <Route path ='/test' element={<TestPage/>} />
-                <Route path ='/letterRead' element={<LetterReadPage/>} />
                 <Route path ='/letterCreate' element={<LetterCreatePage/>} />
                 <Route path = '/ranking' element={<RankingPage/>}/>
                 <Route path = '/tree' element={<LetterTreePage/>}/>
