+++ src/assets/sass/section/_letterRead.scss
@@ -1,5 +1,10 @@
-.letter-read-container {
-        position: relative;
+.tree-background {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background-color: rgba(85, 83, 83, 0.8);
 
     .letter-open-icon {
         height: 325px;
@@ -19,7 +24,7 @@
         .letter {
             width: 341px;
             height: 458.94px;
-            background-color: rgb(255, 127, 113, 0.3);
+            background-color: #F9D3D2;
             border-radius: 7px;
             
             .letter-content {
@@ -38,6 +43,7 @@
             .letter-bottom {
                 padding: 20px;
                 width: 100%;
+                display: flex;
 
                 img {
                     width: 87.91px;
